[REFT] sorting
