+++ src/data_structures/doubly_linked_list.cpp
@@ -216,91 +216,3 @@ void freeDLL(NodeDLL *&head, NodeDLL *&tail)
     }
     tail = NULL;
 }
-
-// --- MergeSort cho danh sách liên kết đôi ---
-// Hàm tách một node khỏi danh sách liên kết đôi
-void splitDLL(NodeDLL *head, NodeDLL **firstHalf, NodeDLL **secondHalf)
-{
-    NodeDLL *slow = head;
-    NodeDLL *fast = head->next;
-
-    // 'fast' di chuyển hai bước, 'slow' di chuyển một bước
-    while (fast != NULL)
-    {
-        fast = fast->next;
-        if (fast != NULL)
-        {
-            slow = slow->next;
-            fast = fast->next;
-        }
-    }
-
-    // 'slow' hiện là ở giữa danh sách
-    // Chia danh sách thành hai phần
-    *firstHalf = head;
-    *secondHalf = slow->next;
-    (*secondHalf)->prev = NULL; // Cập nhật con trỏ prev của nửa thứ hai
-    slow->next = NULL;          // Ngắt liên kết giữa hai nửa
-}
-
-// Hợp nhất hai danh sách liên kết đôi đã sắp xếp
-NodeDLL *mergeDLL(NodeDLL *firstHalf, NodeDLL *secondHalf)
-{
-    // Nếu một trong hai nửa rỗng
-    if (firstHalf == NULL)
-    {
-        return secondHalf;
-    }
-    if (secondHalf == NULL)
-    {
-        return firstHalf;
-    }
-
-    // Chọn node nhỏ hơn làm đầu danh sách kết quả
-    NodeDLL *result = NULL;
-
-    if (firstHalf->info.score <= secondHalf->info.score)
-    {
-        result = firstHalf;
-        result->next = mergeDLL(firstHalf->next, secondHalf);
-        if (result->next != NULL)
-        {
-            result->next->prev = result;
-        }
-    }
-    else
-    {
-        result = secondHalf;
-        result->next = mergeDLL(firstHalf, secondHalf->next);
-        if (result->next != NULL)
-        {
-            result->next->prev = result;
-        }
-    }
-
-    return result;
-}
-
-// Thuật toán MergeSort cho danh sách liên kết đôi
-void mergeSortDLL(NodeDLL **headRef)
-{
-    NodeDLL *head = *headRef;
-    NodeDLL *firstHalf;
-    NodeDLL *secondHalf;
-
-    // Trường hợp cơ bản
-    if (head == NULL || head->next == NULL)
-    {
-        return;
-    }
-
-    // Chia danh sách thành hai nửa
-    splitDLL(head, &firstHalf, &secondHalf);
-
-    // Sắp xếp đệ quy hai nửa
-    mergeSortDLL(&firstHalf);
-    mergeSortDLL(&secondHalf);
-
-    // Hợp nhất hai nửa đã sắp xếp
-    *headRef = mergeDLL(firstHalf, secondHalf);
-}
